producer/requirements.txt
kafka-python-ng
requests